Refactor reservation-related types and improve data handling in reservation modals
- Updated import paths for ReservationFormData to use centralized type definitions.
- Enhanced data handling in ReservationsPage by refining source type and renaming appliedDiscount to extraDiscount for clarity.
- Converted floor values to numbers in ViewReservationModal for consistent data types.
- Simplified button click handling in PassengerAccordion for better readability.
